chore: rename generic save / confirm modal
