+++ frontend/src/app/components/modal/instances/save-confirm-modal.tsx
@@ -1,0 +1,47 @@
+import { FC } from "react";
+import { Modal, Button } from "react-bootstrap";
+import { useAppSelector } from "@hooks/hooks";
+import { selectModalData } from "@store/reducers/app";
+
+type SaveConfirmProps = {
+  close: () => void;
+  submit: () => void;
+};
+
+export const SaveConfirmModal: FC<SaveConfirmProps> = ({ close, submit }) => {
+  const modalData = useAppSelector(selectModalData);
+
+  const { title, description } = modalData;
+
+  const handleConfirm = () => {
+    submit();
+    close();
+  };
+
+  return (
+    <>
+      {title && (
+        <Modal.Header closeButton={true}>
+          <Modal.Title>{title}</Modal.Title>
+        </Modal.Header>
+      )}
+      <Modal.Body>
+        <p>{description}</p>
+      </Modal.Body>
+      <Modal.Footer>
+        <Button
+          variant="outline-primary"
+          onClick={close}
+        >
+          Cancel
+        </Button>
+        <Button
+          variant="primary"
+          onClick={handleConfirm}
+        >
+          Save and close
+        </Button>
+      </Modal.Footer>
+    </>
+  );
+};
