+++ frontend/src/app/components/modal/model-components.ts
@@ -44,7 +44,7 @@ import { MapModal } from "@/app/components/modal/instances/map-modal";
 import { LinkComplaintModal } from "./instances/link-complaint-modal";
 import { AddComplaintToCaseModal } from "./instances/add-complaint-to-case";
 import { CreateAddCaseModal } from "@/app/components/modal/instances/create-add-case";
-import { RemoveActivityFromCaseModal } from "@/app/components/modal/instances/remove-activity-from-case";
+import { SaveConfirmModal } from "@/app/components/modal/instances/save-confirm-modal";
 import { AddEditPartyModal } from "@/app/components/modal/instances/add-party";
 import { CancelConfirmFileUpdateModal } from "./instances/cancel-confirm-file-update-modal";
 import { AddEditTaskAttachmentModal } from "@/app/components/modal/instances/add-task-attachment";
@@ -70,7 +70,7 @@ export const MODAL_COMPONENTS: { [key: string]: React.ComponentType<any> } = {
   [LINK_COMPLAINT]: LinkComplaintModal,
   [ADD_COMPLAINT_TO_CASE]: AddComplaintToCaseModal,
   [CREATE_ADD_CASE]: CreateAddCaseModal,
-  [REMOVE_ACTIVITY_FROM_CASE]: RemoveActivityFromCaseModal,
+  [REMOVE_ACTIVITY_FROM_CASE]: SaveConfirmModal,
   [ADD_PARTY]: AddEditPartyModal,
   [REMOVE_PARTY]: RemovePartyModal,
   [CANCEL_CONFIRM_FILE_UPDATE]: CancelConfirmFileUpdateModal,
